716: cleaned up readme and moved app start to root
