+++ README.md
@@ -1,7 +1,6 @@
 # autocannon-ui
 
 [![ci](https://github.com/nearform/autocannon-ui/actions/workflows/ci.yml/badge.svg)](https://github.com/nearform/autocannon-ui/actions/workflows/ci.yml)
-[![lerna](https://img.shields.io/badge/maintained%20with-lerna-cc00ff.svg)](https://lerna.js.org/)
 [![Conventional Commits](https://img.shields.io/badge/Conventional%20Commits-1.0.0-yellow.svg)](https://conventionalcommits.org)
 
 This monorepo contains packages and applications which provide a GUI for using autocannon while maintaining the same user experience. 
@@ -32,25 +31,27 @@ For understanding the result, please see the `autocannon-compare` [documentation
 
 ## Getting started
 
-- `npm install && npm install -g`
+- `npm install`
 
 The easiest way to try this out is to run:
 
 - `npx autocannon-ui`
-- browse to [`http://localhost:3000`](http://localhost:3000)
+- browse to localhost -- the port is displayed when running this command
 
 ## Development
 
 To easily develop the packages of this repo you can execute:
 
 ```sh
-npm start (npx autocannon-ui after npm install -g)
+npm install
+npm start
 ```
 
 This will run:
 
 - `autocannon-frontend` build in watch mode so you can change the React components and see the result immediately
 - `autocannon-backend` in standalone mode to have a server running
+- the port will be displayed where the application is running
 
 ### Designs
 
@@ -66,17 +67,13 @@ A library of React components which provide the UI.
 
 A standalone Fastify application.
 
-### `autocannon-ui`
-
-Web wrapper for running both frontend and backend. (root level)
 
 ## Running standalone application
 
 The application can be installed as a global npm package and run as a standalone tool from command line.
 
-To install it and to run web application:
+To install it and to run the application:
 
 ```sh
-npm install
 npx autocannon-ui
 ```
