+++ cli.js
@@ -1,3 +1,3 @@
 #!/usr/bin/env node
 
-import './packages/autocannon-ui-backend/app.js'
+import './app.js'
