--- packages/autocannon-ui-backend/README.md
@@ -1,7 +1,0 @@
-#autocannon-ui-backend
-
-This package contains [Fastify](https://github.com/fastify/fastify) backend API required by [autocannon-ui](https://github.com/nearform/autocannon-ui) application.
-
-##Development
-
-Run `npm start` to start a Fastify server which serves the API.
